refactor(inventory): drop python_ prefix on Python-only modules
Review #8 on PR #794. The arc's language qualifiers were
inconsistent: python_callgraph / python_interproc /
python_taint_summaries carried a python_ prefix while cfg_builder /
cfg_builder_cpp used a _cpp suffix on the second-language variant.
Adopt the cfg_builder convention — no suffix is the Python default,
suffix only when a second-language variant exists:

  python_callgraph        → callgraph
  python_interproc        → interproc
  python_taint_summaries  → taint_summaries

(plus the matching test files). Update all imports and doc/docstring
references. Note: callgraph.py is distinct from the pre-existing
call_graph.py — different subsystem (module-local sanitizer-cut call
graph vs the inventory call-graph builder).
