+++ core/inventory/callgraph.py
